+++ Services/Frs/V2/Model/Attributes.php
@@ -22,18 +22,10 @@ class Attributes implements ModelInterface, ArrayAccess
     * Array of property to type mappings. Used for (de)serialization
     * dress  dress
     * glass  是否带眼镜： • yes：带眼镜 • dark：带墨镜 • none：未戴眼镜 • unknown：未知
-    * gender  性别： • male：男 • female：女 • unknown：未知
-    * yawAngle  围绕Y轴旋转，偏航角，范围[-180,180]。
-    * rollAngle  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    * pitchAngle  围绕X轴旋转，俯仰角，范围[-180,180]。
     * hat  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
-    * headpose  人脸轮廓坐标值。
     * age  年龄。
-    * smile  笑脸。
     * mask  是否戴口罩： • yes：戴口罩 • none：未戴口罩 • unknown：未知
     * beard  胡须： • yes：有胡须 • none：无胡须 • unknown：未知
-    * skin  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    * ethnic  民族： • han：汉族 • other：其他 • unknown：未知
     * phototype  图片类型： • idcard：证件照 • monitor：摄像头监控 • internet photo：网络图片
     * quality  quality
     * hair  发型： • long：长发 • short：短发 • unknown：未知
@@ -45,18 +37,10 @@ class Attributes implements ModelInterface, ArrayAccess
     protected static $openAPITypes = [
             'dress' => '\HuaweiCloud\SDK\Frs\V2\Model\Dress',
             'glass' => 'string',
-            'gender' => 'string',
-            'yawAngle' => 'double',
-            'rollAngle' => 'double',
-            'pitchAngle' => 'double',
             'hat' => 'string',
-            'headpose' => 'double[]',
             'age' => 'int',
-            'smile' => 'string',
             'mask' => 'string',
             'beard' => 'string',
-            'skin' => 'string',
-            'ethnic' => 'string',
             'phototype' => 'string',
             'quality' => '\HuaweiCloud\SDK\Frs\V2\Model\FaceQuality',
             'hair' => 'string',
@@ -68,18 +52,10 @@ class Attributes implements ModelInterface, ArrayAccess
     * Array of property to format mappings. Used for (de)serialization
     * dress  dress
     * glass  是否带眼镜： • yes：带眼镜 • dark：带墨镜 • none：未戴眼镜 • unknown：未知
-    * gender  性别： • male：男 • female：女 • unknown：未知
-    * yawAngle  围绕Y轴旋转，偏航角，范围[-180,180]。
-    * rollAngle  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    * pitchAngle  围绕X轴旋转，俯仰角，范围[-180,180]。
     * hat  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
-    * headpose  人脸轮廓坐标值。
     * age  年龄。
-    * smile  笑脸。
     * mask  是否戴口罩： • yes：戴口罩 • none：未戴口罩 • unknown：未知
     * beard  胡须： • yes：有胡须 • none：无胡须 • unknown：未知
-    * skin  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    * ethnic  民族： • han：汉族 • other：其他 • unknown：未知
     * phototype  图片类型： • idcard：证件照 • monitor：摄像头监控 • internet photo：网络图片
     * quality  quality
     * hair  发型： • long：长发 • short：短发 • unknown：未知
@@ -91,18 +67,10 @@ class Attributes implements ModelInterface, ArrayAccess
     protected static $openAPIFormats = [
         'dress' => null,
         'glass' => null,
-        'gender' => null,
-        'yawAngle' => 'double',
-        'rollAngle' => 'double',
-        'pitchAngle' => 'double',
         'hat' => null,
-        'headpose' => 'double',
         'age' => 'int32',
-        'smile' => null,
         'mask' => null,
         'beard' => null,
-        'skin' => null,
-        'ethnic' => null,
         'phototype' => null,
         'quality' => null,
         'hair' => null,
@@ -135,18 +103,10 @@ class Attributes implements ModelInterface, ArrayAccess
     * and the value is the original name
     * dress  dress
     * glass  是否带眼镜： • yes：带眼镜 • dark：带墨镜 • none：未戴眼镜 • unknown：未知
-    * gender  性别： • male：男 • female：女 • unknown：未知
-    * yawAngle  围绕Y轴旋转，偏航角，范围[-180,180]。
-    * rollAngle  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    * pitchAngle  围绕X轴旋转，俯仰角，范围[-180,180]。
     * hat  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
-    * headpose  人脸轮廓坐标值。
     * age  年龄。
-    * smile  笑脸。
     * mask  是否戴口罩： • yes：戴口罩 • none：未戴口罩 • unknown：未知
     * beard  胡须： • yes：有胡须 • none：无胡须 • unknown：未知
-    * skin  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    * ethnic  民族： • han：汉族 • other：其他 • unknown：未知
     * phototype  图片类型： • idcard：证件照 • monitor：摄像头监控 • internet photo：网络图片
     * quality  quality
     * hair  发型： • long：长发 • short：短发 • unknown：未知
@@ -158,18 +118,10 @@ class Attributes implements ModelInterface, ArrayAccess
     protected static $attributeMap = [
             'dress' => 'dress',
             'glass' => 'glass',
-            'gender' => 'gender',
-            'yawAngle' => 'yaw_angle',
-            'rollAngle' => 'roll_angle',
-            'pitchAngle' => 'pitch_angle',
             'hat' => 'hat',
-            'headpose' => 'headpose',
             'age' => 'age',
-            'smile' => 'smile',
             'mask' => 'mask',
             'beard' => 'beard',
-            'skin' => 'skin',
-            'ethnic' => 'ethnic',
             'phototype' => 'phototype',
             'quality' => 'quality',
             'hair' => 'hair',
@@ -181,18 +133,10 @@ class Attributes implements ModelInterface, ArrayAccess
     * Array of attributes to setter functions (for deserialization of responses)
     * dress  dress
     * glass  是否带眼镜： • yes：带眼镜 • dark：带墨镜 • none：未戴眼镜 • unknown：未知
-    * gender  性别： • male：男 • female：女 • unknown：未知
-    * yawAngle  围绕Y轴旋转，偏航角，范围[-180,180]。
-    * rollAngle  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    * pitchAngle  围绕X轴旋转，俯仰角，范围[-180,180]。
     * hat  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
-    * headpose  人脸轮廓坐标值。
     * age  年龄。
-    * smile  笑脸。
     * mask  是否戴口罩： • yes：戴口罩 • none：未戴口罩 • unknown：未知
     * beard  胡须： • yes：有胡须 • none：无胡须 • unknown：未知
-    * skin  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    * ethnic  民族： • han：汉族 • other：其他 • unknown：未知
     * phototype  图片类型： • idcard：证件照 • monitor：摄像头监控 • internet photo：网络图片
     * quality  quality
     * hair  发型： • long：长发 • short：短发 • unknown：未知
@@ -204,18 +148,10 @@ class Attributes implements ModelInterface, ArrayAccess
     protected static $setters = [
             'dress' => 'setDress',
             'glass' => 'setGlass',
-            'gender' => 'setGender',
-            'yawAngle' => 'setYawAngle',
-            'rollAngle' => 'setRollAngle',
-            'pitchAngle' => 'setPitchAngle',
             'hat' => 'setHat',
-            'headpose' => 'setHeadpose',
             'age' => 'setAge',
-            'smile' => 'setSmile',
             'mask' => 'setMask',
             'beard' => 'setBeard',
-            'skin' => 'setSkin',
-            'ethnic' => 'setEthnic',
             'phototype' => 'setPhototype',
             'quality' => 'setQuality',
             'hair' => 'setHair',
@@ -227,18 +163,10 @@ class Attributes implements ModelInterface, ArrayAccess
     * Array of attributes to getter functions (for serialization of requests)
     * dress  dress
     * glass  是否带眼镜： • yes：带眼镜 • dark：带墨镜 • none：未戴眼镜 • unknown：未知
-    * gender  性别： • male：男 • female：女 • unknown：未知
-    * yawAngle  围绕Y轴旋转，偏航角，范围[-180,180]。
-    * rollAngle  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    * pitchAngle  围绕X轴旋转，俯仰角，范围[-180,180]。
     * hat  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
-    * headpose  人脸轮廓坐标值。
     * age  年龄。
-    * smile  笑脸。
     * mask  是否戴口罩： • yes：戴口罩 • none：未戴口罩 • unknown：未知
     * beard  胡须： • yes：有胡须 • none：无胡须 • unknown：未知
-    * skin  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    * ethnic  民族： • han：汉族 • other：其他 • unknown：未知
     * phototype  图片类型： • idcard：证件照 • monitor：摄像头监控 • internet photo：网络图片
     * quality  quality
     * hair  发型： • long：长发 • short：短发 • unknown：未知
@@ -250,18 +178,10 @@ class Attributes implements ModelInterface, ArrayAccess
     protected static $getters = [
             'dress' => 'getDress',
             'glass' => 'getGlass',
-            'gender' => 'getGender',
-            'yawAngle' => 'getYawAngle',
-            'rollAngle' => 'getRollAngle',
-            'pitchAngle' => 'getPitchAngle',
             'hat' => 'getHat',
-            'headpose' => 'getHeadpose',
             'age' => 'getAge',
-            'smile' => 'getSmile',
             'mask' => 'getMask',
             'beard' => 'getBeard',
-            'skin' => 'getSkin',
-            'ethnic' => 'getEthnic',
             'phototype' => 'getPhototype',
             'quality' => 'getQuality',
             'hair' => 'getHair',
@@ -329,18 +249,10 @@ class Attributes implements ModelInterface, ArrayAccess
     {
         $this->container['dress'] = isset($data['dress']) ? $data['dress'] : null;
         $this->container['glass'] = isset($data['glass']) ? $data['glass'] : null;
-        $this->container['gender'] = isset($data['gender']) ? $data['gender'] : null;
-        $this->container['yawAngle'] = isset($data['yawAngle']) ? $data['yawAngle'] : null;
-        $this->container['rollAngle'] = isset($data['rollAngle']) ? $data['rollAngle'] : null;
-        $this->container['pitchAngle'] = isset($data['pitchAngle']) ? $data['pitchAngle'] : null;
         $this->container['hat'] = isset($data['hat']) ? $data['hat'] : null;
-        $this->container['headpose'] = isset($data['headpose']) ? $data['headpose'] : null;
         $this->container['age'] = isset($data['age']) ? $data['age'] : null;
-        $this->container['smile'] = isset($data['smile']) ? $data['smile'] : null;
         $this->container['mask'] = isset($data['mask']) ? $data['mask'] : null;
         $this->container['beard'] = isset($data['beard']) ? $data['beard'] : null;
-        $this->container['skin'] = isset($data['skin']) ? $data['skin'] : null;
-        $this->container['ethnic'] = isset($data['ethnic']) ? $data['ethnic'] : null;
         $this->container['phototype'] = isset($data['phototype']) ? $data['phototype'] : null;
         $this->container['quality'] = isset($data['quality']) ? $data['quality'] : null;
         $this->container['hair'] = isset($data['hair']) ? $data['hair'] : null;
@@ -362,41 +274,17 @@ class Attributes implements ModelInterface, ArrayAccess
         if ($this->container['glass'] === null) {
             $invalidProperties[] = "'glass' can't be null";
         }
-        if ($this->container['gender'] === null) {
-            $invalidProperties[] = "'gender' can't be null";
-        }
-        if ($this->container['yawAngle'] === null) {
-            $invalidProperties[] = "'yawAngle' can't be null";
-        }
-        if ($this->container['rollAngle'] === null) {
-            $invalidProperties[] = "'rollAngle' can't be null";
-        }
-        if ($this->container['pitchAngle'] === null) {
-            $invalidProperties[] = "'pitchAngle' can't be null";
-        }
         if ($this->container['hat'] === null) {
             $invalidProperties[] = "'hat' can't be null";
         }
-        if ($this->container['headpose'] === null) {
-            $invalidProperties[] = "'headpose' can't be null";
-        }
         if ($this->container['age'] === null) {
             $invalidProperties[] = "'age' can't be null";
-        }
-        if ($this->container['smile'] === null) {
-            $invalidProperties[] = "'smile' can't be null";
         }
         if ($this->container['mask'] === null) {
             $invalidProperties[] = "'mask' can't be null";
         }
         if ($this->container['beard'] === null) {
             $invalidProperties[] = "'beard' can't be null";
-        }
-        if ($this->container['skin'] === null) {
-            $invalidProperties[] = "'skin' can't be null";
-        }
-        if ($this->container['ethnic'] === null) {
-            $invalidProperties[] = "'ethnic' can't be null";
         }
         if ($this->container['phototype'] === null) {
             $invalidProperties[] = "'phototype' can't be null";
@@ -476,102 +364,6 @@ class Attributes implements ModelInterface, ArrayAccess
     }
 
     /**
-    * Gets gender
-    *  性别： • male：男 • female：女 • unknown：未知
-    *
-    * @return string
-    */
-    public function getGender()
-    {
-        return $this->container['gender'];
-    }
-
-    /**
-    * Sets gender
-    *
-    * @param string $gender 性别： • male：男 • female：女 • unknown：未知
-    *
-    * @return $this
-    */
-    public function setGender($gender)
-    {
-        $this->container['gender'] = $gender;
-        return $this;
-    }
-
-    /**
-    * Gets yawAngle
-    *  围绕Y轴旋转，偏航角，范围[-180,180]。
-    *
-    * @return double
-    */
-    public function getYawAngle()
-    {
-        return $this->container['yawAngle'];
-    }
-
-    /**
-    * Sets yawAngle
-    *
-    * @param double $yawAngle 围绕Y轴旋转，偏航角，范围[-180,180]。
-    *
-    * @return $this
-    */
-    public function setYawAngle($yawAngle)
-    {
-        $this->container['yawAngle'] = $yawAngle;
-        return $this;
-    }
-
-    /**
-    * Gets rollAngle
-    *  围绕Z轴旋转，翻滚角，范围[-180,180]。
-    *
-    * @return double
-    */
-    public function getRollAngle()
-    {
-        return $this->container['rollAngle'];
-    }
-
-    /**
-    * Sets rollAngle
-    *
-    * @param double $rollAngle 围绕Z轴旋转，翻滚角，范围[-180,180]。
-    *
-    * @return $this
-    */
-    public function setRollAngle($rollAngle)
-    {
-        $this->container['rollAngle'] = $rollAngle;
-        return $this;
-    }
-
-    /**
-    * Gets pitchAngle
-    *  围绕X轴旋转，俯仰角，范围[-180,180]。
-    *
-    * @return double
-    */
-    public function getPitchAngle()
-    {
-        return $this->container['pitchAngle'];
-    }
-
-    /**
-    * Sets pitchAngle
-    *
-    * @param double $pitchAngle 围绕X轴旋转，俯仰角，范围[-180,180]。
-    *
-    * @return $this
-    */
-    public function setPitchAngle($pitchAngle)
-    {
-        $this->container['pitchAngle'] = $pitchAngle;
-        return $this;
-    }
-
-    /**
     * Gets hat
     *  是否戴帽子： • yes：戴帽子 • none：未戴帽子 • unknown：未知
     *
@@ -596,30 +388,6 @@ class Attributes implements ModelInterface, ArrayAccess
     }
 
     /**
-    * Gets headpose
-    *  人脸轮廓坐标值。
-    *
-    * @return double[]
-    */
-    public function getHeadpose()
-    {
-        return $this->container['headpose'];
-    }
-
-    /**
-    * Sets headpose
-    *
-    * @param double[] $headpose 人脸轮廓坐标值。
-    *
-    * @return $this
-    */
-    public function setHeadpose($headpose)
-    {
-        $this->container['headpose'] = $headpose;
-        return $this;
-    }
-
-    /**
     * Gets age
     *  年龄。
     *
@@ -640,30 +408,6 @@ class Attributes implements ModelInterface, ArrayAccess
     public function setAge($age)
     {
         $this->container['age'] = $age;
-        return $this;
-    }
-
-    /**
-    * Gets smile
-    *  笑脸。
-    *
-    * @return string
-    */
-    public function getSmile()
-    {
-        return $this->container['smile'];
-    }
-
-    /**
-    * Sets smile
-    *
-    * @param string $smile 笑脸。
-    *
-    * @return $this
-    */
-    public function setSmile($smile)
-    {
-        $this->container['smile'] = $smile;
         return $this;
     }
 
@@ -712,54 +456,6 @@ class Attributes implements ModelInterface, ArrayAccess
     public function setBeard($beard)
     {
         $this->container['beard'] = $beard;
-        return $this;
-    }
-
-    /**
-    * Gets skin
-    *  肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    *
-    * @return string
-    */
-    public function getSkin()
-    {
-        return $this->container['skin'];
-    }
-
-    /**
-    * Sets skin
-    *
-    * @param string $skin 肤色： • brown：棕 • yellow：黄 • white：白 • black：黑 • unknown：未知
-    *
-    * @return $this
-    */
-    public function setSkin($skin)
-    {
-        $this->container['skin'] = $skin;
-        return $this;
-    }
-
-    /**
-    * Gets ethnic
-    *  民族： • han：汉族 • other：其他 • unknown：未知
-    *
-    * @return string
-    */
-    public function getEthnic()
-    {
-        return $this->container['ethnic'];
-    }
-
-    /**
-    * Sets ethnic
-    *
-    * @param string $ethnic 民族： • han：汉族 • other：其他 • unknown：未知
-    *
-    * @return $this
-    */
-    public function setEthnic($ethnic)
-    {
-        $this->container['ethnic'] = $ethnic;
         return $this;
     }